\name{kkma.extractPOS}
\alias{kkma.extractPOS}
\title{extractPOS using Kkma}
\description{
KKMA
}
\usage{
kkma.extractPOS(strings, POS = "NV")
kkma.extractPOS(strings, POS = "N")
kkma.extractPOS(strings, POS = "V")
}
\arguments{
  \item{strings}{input characters}
  \item{POS}{input character}
}
\value{
N : Noun
V : Verb
NV : N or V
}
