\name{KoNLPQ-package}
\alias{KoNLPQ-package}
\alias{KoNLPQ}
\docType{package}
\title{
\packageTitle{KoNLPQ}
}
\description{
\packageDescription{KoNLPQ}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KoNLPQ}
\packageIndices{KoNLPQ}
An overview of how to use the package, including the most
important functions
}
\author{
\packageAuthor{KoNLPQ}

Maintainer: \packageMaintainer{KoNLPQ}
}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from file ~~
KEYWORDS in the R documentation directory ~~